requirements.txt
alembic==1.9.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
colorama==0.4.6
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.46
tornado==6.1
tzdata==2022.7
tzlocal==4.2
